TESTED.  No large problems seen.
Good for all three targets, as-is.

UART support is probably reinstated now (did not
check; most likely not in this branch's history,
as once the 'correct' problem was identified, the
removal of UART related code was abandoned, and
a new effort was made, here, which did not touch
upon UART issues at all).

This object:

   compile_time_choice.S.obj

Seems to be the new means to build QSPI flashROM support.

The 'generic' flash driver is no longer appropriate,
or at least has less of a use case, with the recent,
upstream improvement, of the pico-sdk - the define
for generic flash support isn't used, now, in the
script used to cmake.

   cmake $   make

  Scanning dependencies of target bs2_default
  Building ASM object pico-sdk/src/rp2_common/boot_stage2/CMakeFiles/bs2_default.dir/compile_time_choice.S.obj

	modified:   forth.c

On branch dvlp-aa-all-a-